♻️ refactor : Moving interface file locations and modifying code based on export rules
- Separate dashboard interfaces
- Export each interface as default, specifying interface as the export type in the index.ts file

Related issue: #68
