+++ src/app/components/card/Invitation.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import Image from 'next/image';
-import {InvitationProps} from '@/src/interfaces/group/invitationProps';
+import {InvitationProps} from '@/src/interfaces/group';
 
 const Invitation = ({ImageSrc, GroupName, ManagerName}: InvitationProps) => {
   return (
