+++ src/components/dashboard/TabsMapping.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {TabsMappingProps} from '@/src/interfaces/group';
+import {TabsMappingProps} from '@/src/interfaces/dashboard';
 
 const TabsMapping = ({activeTab, tabMapping}: TabsMappingProps) => {
   return <>{tabMapping[activeTab]}</>;
