+++ src/interfaces/dashboard/index.ts
@@ -1,0 +1,4 @@
+import DashboardProps from './dashboardProps';
+import TabsMappingProps from './tabsMappingProps';
+
+export type {DashboardProps, TabsMappingProps};
